Remove "ps"-based OIDC Discovery Provider readiness probes
We are migrating away from the alpine images towards the scratch images
being the default. In a scratch image, we don't have the "ps" binary.

There is a bug in the OIDC Discovery Provider that prevents the HTTP
liveness/readiness endpoint from being available outside the container
(see spiffe#spire/3629), so just remove the readiness probes for now.
Turn on the health check endpoint to the OIDC Discovery Provider so that
probes can be added later on once the issue is resolved.

Also update some old configs that are no longer correct to get the
examples to run properly.

Signed-off-by: Ryan Turner <turner@uber.com>