k8s/quickstart: specify default namespace for client deployment
The registration entry for the workload specifies a selector of
`k8s:ns:default`, which means the workload should be running in the
default namespace. If the namespace isn't specified in the yaml, the
workload will deploy to whichever namespace your context is currently
set to, which can cause issue spiffe/spire-tutorials#123.

Fixes #123.